修复软删除model移入回收站会删除已上传文件问题

+++ src/Repositories/EloquentRepository.php
@@ -467,9 +467,10 @@ class EloquentRepository extends Repository implements TreeRepository
                 $model->forceDelete();
 
                 return;
+            } elseif (! $this->isSoftDeletes) {
+                $form->deleteFiles($data);
             }
 
-            $form->deleteFiles($data);
             $model->delete();
         });
